requirements.txt
aio-pika==9.4.3
aioconsole==0.8.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiomonitor==0.7.0
aiormq==6.8.1
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
janus==1.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
prompt_toolkit==3.0.48
psutil==6.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-logging-loki==0.3.1
PyYAML==6.0.2
requests==2.32.3
rfc3339==6.2
schedule==1.2.2
setuptools==75.1.0
terminaltables==3.1.10
trafaret==2.1.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.12.1
zope.interface==7.0.3
